+++ CHANGES.md
@@ -1,6 +1,12 @@
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
 <!-- markdownlint-disable-next-line first-line-h1 -->
+## Unreleased
+
+### Fixed
+
+* vertigo-cli: Workaround for WASM instantiation on different rust versions.
+
 ## 0.8.2 - 2025-09-10
 
 ### Added